Refactor/rename DAOs etc.
