+++ src/test/scala/uk/gov/ons/sbr/data/model/CompanyDaoTest.scala
@@ -2,10 +2,9 @@ package uk.gov.ons.sbr.data.model
 
 import com.typesafe.config.ConfigFactory
 import org.scalatest.{BeforeAndAfterAll, BeforeAndAfterEach, FlatSpec, Matchers}
-import uk.gov.ons.sbr.data.SbrDatabase
-import uk.gov.ons.sbr.data.utils.DbSchemaService
+import uk.gov.ons.sbr.data.db.{DbSchemaService, SbrDatabase}
 
-class PayeRepoTest extends FlatSpec with BeforeAndAfterAll with BeforeAndAfterEach with Matchers {
+class CompanyDaoTest extends FlatSpec with BeforeAndAfterAll with BeforeAndAfterEach with Matchers {
 
   // Set up DB...
   val config = ConfigFactory.load()
@@ -15,11 +14,10 @@ class PayeRepoTest extends FlatSpec with BeforeAndAfterAll with BeforeAndAfterEa
   // Get implicit session for voodoo with DB operations below
   implicit val session = db.session
 
-  // Use same entity Repo object for all tests
-  val entRepo = EnterpriseRepo
-  val unitRepo = LegalUnitRepo
-  val chRepo = CompanyRepo
-  val payeRepo = PayeRepo
+  // Use same Repo object for all tests
+  val entRepo = EnterpriseDao
+  val unitRepo = LegalUnitDao
+  val chRepo = CompanyDao
 
   override def beforeAll(): Unit = {
     super.beforeAll()
@@ -38,8 +36,6 @@ class PayeRepoTest extends FlatSpec with BeforeAndAfterAll with BeforeAndAfterEa
 
   // delete all data between tests
   override def beforeEach(): Unit = {
-    payeRepo.deleteAll
-
     chRepo.deleteAll
 
     unitRepo.deleteAll
@@ -50,10 +46,9 @@ class PayeRepoTest extends FlatSpec with BeforeAndAfterAll with BeforeAndAfterEa
 
   /** * TESTS START HERE ***/
 
+  behavior of "CompanyRepo"
 
-  behavior of "PayeRepo"
-
-  it should "insert new Enterprise, Legal Unit and PAYE, then query PAYE correctly with Ref Period" in {
+  it should "insert new Enterprise, Legal Unit and Company, then query Company correctly with Ref Period" in {
 
     val entref = 100L
     val refperiod = 201708 // not default period
@@ -67,20 +62,19 @@ class PayeRepoTest extends FlatSpec with BeforeAndAfterAll with BeforeAndAfterEa
     val leu: LegalUnit = LegalUnit(ref_period = refperiod, ubrn = ubrn, entref = entref, businessname = Option(s"Legal Unit $ubrn"))
     val newLeu = unitRepo.insert(leu)
 
-    // Now create PAYE for this LEU
-    val pref = "PAYE0001"
-    val paye: Paye = Paye(ref_period = refperiod, payeref = pref, ubrn = ubrn)
-    val newPaye = payeRepo.insert(paye)
+    // Now create Company for this LEU
+    val coNo = "COMPANY0001"
+    val co: Company = Company(ref_period = refperiod, companynumber = coNo, companyname = Some(s"Test company $coNo"), ubrn = ubrn)
+    val newCo = chRepo.insert(co)
 
     // Now see if we can query it back
-    val fetched = payeRepo.getPaye(refperiod, pref)
+    val fetched = chRepo.getCompany(refperiod, coNo)
 
-    fetched shouldBe (Some(newPaye))
+    fetched shouldBe (Some(newCo))
 
   }
 
-
-  it should "insert new Ent/LEU/PAYE then delete PAYE correctly" in {
+  it should "insert new Ent/LEU and delete Company correctly" in {
 
     val entref = 100L
     val refperiod = 201708 // not default period
@@ -94,23 +88,22 @@ class PayeRepoTest extends FlatSpec with BeforeAndAfterAll with BeforeAndAfterEa
     val leu: LegalUnit = LegalUnit(ref_period = refperiod, ubrn = ubrn, entref = entref, businessname = Option(s"Legal Unit $ubrn"))
     val newLeu = unitRepo.insert(leu)
 
-    // Now create PAYE for this LEU
-    val pref = "PAYE0001"
-    val paye: Paye = Paye(ref_period = refperiod, payeref = pref, ubrn = ubrn)
-    val newPaye = payeRepo.insert(paye)
+    // Now create Company for this LEU
+    val coNo = "COMPANY0001"
+    val co: Company = Company(ref_period = refperiod, companynumber = coNo, companyname = Some(s"Test company $coNo"), ubrn = ubrn)
+    val newCo = chRepo.insert(co)
 
-
-    // delete new PAYE
-    newPaye.destroy()
+    // delete new Company
+    newCo.destroy()
 
     // Now see if we can query it back  (should be gone)
-    val fetched = payeRepo.getPaye(refperiod, pref)
+    val fetched = chRepo.getCompany(refperiod, coNo)
 
     fetched shouldBe (None)
 
   }
 
-  it should "query right number of PAYE records for a given Legal Unit" in {
+  it should "query right number of Company records for a given Legal Unit" in {
 
     val entref = 101L
     val refperiod = 201708 // not default period
@@ -124,23 +117,23 @@ class PayeRepoTest extends FlatSpec with BeforeAndAfterAll with BeforeAndAfterEa
     val leu: LegalUnit = LegalUnit(ref_period = refperiod, ubrn = ubrn, entref = entref, businessname = Option(s"Legal Unit $ubrn"))
     val newLeu = unitRepo.insert(leu)
 
-    // create several records for this LEU
+    // create several Companies for this LEU
     val numUnits = 10
     val ids = (1 to numUnits)
     ids foreach { id =>
-      val pref = s"PAYE$id"
-      val paye: Paye = Paye(ref_period = refperiod, payeref = pref, ubrn = ubrn)
-      val newPaye = payeRepo.insert(paye)
+      val coNo = s"COMPANY$id"
+      val data: Company = Company(ref_period = refperiod, companynumber = coNo, companyname = Some(s"Test company $coNo"), ubrn = ubrn)
+      chRepo.insert(data)
     }
 
     // count the records for this legal unit
-    val fetched = payeRepo.getPayesForLegalUnit(refperiod, ubrn)
+    val fetched = chRepo.getCompaniesForLegalUnit(refperiod, ubrn)
 
     // Check we got the right number of records
     fetched.size shouldBe numUnits
   }
 
-  it should "query specific PAYE correctly from several records" in {
+  it should "query specific Company correctly from several records" in {
     val entref = 101L
     val refperiod = 201708 // not default period
 
@@ -153,34 +146,33 @@ class PayeRepoTest extends FlatSpec with BeforeAndAfterAll with BeforeAndAfterEa
     val leu: LegalUnit = LegalUnit(ref_period = refperiod, ubrn = ubrn, entref = entref, businessname = Option(s"Legal Unit $ubrn"))
     val newLeu = unitRepo.insert(leu)
 
-    // create several records for this LEU
+    // create several Companies for this LEU
     val numUnits = 10
     val ids = (1 to numUnits)
     ids foreach { id =>
-      val pref = s"PAYE$id"
-      val paye: Paye = Paye(ref_period = refperiod, payeref = pref, name1 = Some(s"PAYE-NAME-$pref"),ubrn = ubrn)
-      val newPaye = payeRepo.insert(paye)
+      val coNo = s"COMPANY$id"
+      val data: Company = Company(ref_period = refperiod, companynumber = coNo, companyname = Some(s"Test company $coNo"), ubrn = ubrn)
+      chRepo.insert(data)
     }
 
     // query a random record back for an ID between 1 and num (watch out for 0 coming back from random gen)
     val r = scala.util.Random
     val qid: String = r.nextInt(numUnits) match {
-      case 0 => s"PAYE1"
-      case n => s"PAYE$n"
+      case 0 => s"COMPANY1"
+      case n => s"COMPANY$n"
     }
 
     // fetch the data
-    val fetched = payeRepo.getPaye(refperiod, qid)
+    val fetched = chRepo.getCompany(refperiod, qid)
 
-    val expectedName = Some(s"PAYE-NAME-$qid")
+    val expectedName = Some(s"Test company $qid")
     // Remember name is an option as well
-    val actualName = fetched.map(_.name1).flatten
+    val actualName = fetched.map(_.companyname).flatten
 
     actualName shouldBe expectedName
   }
 
-
-  it should "count PAYE records correctly" in {
+  it should "count Company records correctly" in {
 
     val entref = 101L
     val refperiod = 201706
@@ -196,13 +188,13 @@ class PayeRepoTest extends FlatSpec with BeforeAndAfterAll with BeforeAndAfterEa
     val leu1: LegalUnit = LegalUnit(ref_period = refperiod, ubrn = ubrn1, entref = entref, businessname = Option(s"Legal Unit $ubrn1"))
     val newLeu1 = unitRepo.insert(leu1)
 
-    // create several records for this LEU
+    // create several Companies for this LEU
     val numUnits = 10
     val ids = (1 to numUnits)
     ids foreach { id =>
-      val pref = s"PAYE$id"
-      val paye: Paye = Paye(ref_period = refperiod, payeref = pref, ubrn = ubrn1)
-      val newPaye = payeRepo.insert(paye)
+      val coNo = s"COMPANY$id"
+      val data: Company = Company(ref_period = refperiod, companynumber = coNo, companyname = Some(s"Test company $coNo"), ubrn = ubrn1)
+      chRepo.insert(data)
     }
 
     val ubrn2 = 10002L
@@ -210,18 +202,18 @@ class PayeRepoTest extends FlatSpec with BeforeAndAfterAll with BeforeAndAfterEa
     val leu2: LegalUnit = LegalUnit(ref_period = refperiod, ubrn = ubrn2, entref = entref, businessname = Option(s"Legal Unit $ubrn2"))
     val newLeu2 = unitRepo.insert(leu2)
 
-    // create several records for this LEU
+    // create several Companies for this LEU
     val start = numUnits + 1
     val end = numUnits + numUnits
     (start to end) foreach { id =>
-      val pref = s"PAYE$id"
-      val paye: Paye = Paye(ref_period = refperiod, payeref = pref, ubrn = ubrn2)
-      val newPaye = payeRepo.insert(paye)
+      val coNo = s"COMPANY$id"
+      val data: Company = Company(ref_period = refperiod, companynumber = coNo, companyname = Some(s"Test company $coNo"), ubrn = ubrn2)
+      chRepo.insert(data)
     }
 
-    // count ALL records i.e. includes both LEUs
+    // count ALL Companies i.e. includes both LEUs
     val expected = numUnits + numUnits
-    val counted = payeRepo.count()
+    val counted = chRepo.count()
 
     counted shouldBe expected
